mss*: Change type of array stride parameters to ptrdiff_t
ptrdiff_t is the correct type for array strides and similar.

+++ libavcodec/mss2dsp.h
@@ -26,23 +26,25 @@
 #ifndef AVCODEC_MSS2DSP_H
 #define AVCODEC_MSS2DSP_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct MSS2DSPContext {
-    void (*mss2_blit_wmv9)(uint8_t *dst, int dst_stride,
-                           const uint8_t *srcy, int srcy_stride,
+    void (*mss2_blit_wmv9)(uint8_t *dst, ptrdiff_t dst_stride,
+                           const uint8_t *srcy, ptrdiff_t srcy_stride,
                            const uint8_t *srcu, const uint8_t *srcv,
-                           int srcuv_stride, int w, int h);
-    void (*mss2_blit_wmv9_masked)(uint8_t *dst, int dst_stride,
+                           ptrdiff_t srcuv_stride, int w, int h);
+    void (*mss2_blit_wmv9_masked)(uint8_t *dst, ptrdiff_t dst_stride,
                                   int maskcolor, const uint8_t *mask,
-                                  int mask_stride,
-                                  const uint8_t *srcy, int srcy_stride,
+                                  ptrdiff_t mask_stride,
+                                  const uint8_t *srcy, ptrdiff_t srcy_stride,
                                   const uint8_t *srcu, const uint8_t *srcv,
-                                  int srcuv_stride, int w, int h);
-    void (*mss2_gray_fill_masked)(uint8_t *dst, int dst_stride,
+                                  ptrdiff_t srcuv_stride, int w, int h);
+    void (*mss2_gray_fill_masked)(uint8_t *dst, ptrdiff_t dst_stride,
                                   int maskcolor, const uint8_t *mask,
-                                  int mask_stride, int w, int h);
-    void (*upsample_plane)(uint8_t *plane, int plane_stride, int w, int h);
+                                  ptrdiff_t mask_stride, int w, int h);
+    void (*upsample_plane)(uint8_t *plane, ptrdiff_t plane_stride,
+                           int w, int h);
 } MSS2DSPContext;
 
 void ff_mss2dsp_init(MSS2DSPContext *dsp);
